--- src/main/code/de/tu_chemnitz/se/exercise/Greeter.kt
@@ -1,9 +1,0 @@
-package de.tu_chemnitz.se.exercise
-
-class Greeter(
-    private val out: OutputDevice
-) {
-    fun greet(whom: String = "world") {
-        out.printLine("Hello, $whom!")
-    }
-}
