--- src/main/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsInputValidator.kt
@@ -1,6 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-class CodeChartsInputValidator {
-    fun showInputLine() {}
-    fun validateStringInput() {}
-}
