--- src/main/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsMatrix.kt
@@ -1,7 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-class CodeChartsMatrix {
-    fun showMatrix() {}
-    private fun showStrings() {}
-    private fun rasterizeField() {}
-}
