--- src/main/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsPicture.kt
@@ -1,6 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-class CodeChartsPicture {
-    fun showPicture() {}
-    fun hidePicture() {}
-}
