--- src/main/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsStringHandler.kt
@@ -1,14 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-class CodeChartsStringHandler {
-    private lateinit var generatedStrings: Set<String>
-
-    /**
-     * @param Input Number of Chars in the generated strings.
-     * @param AllowedChars ByteArray with 3 modes in the following order: Upper Letters, Lower Letters, Numbers
-     *     where 1 is enabled and 0 is not enabled.
-     * @return StringArray
-     */
-    fun generateStrings(Input: Int, AllowedChars: ByteArray) {}
-    fun getStrings() {}
-}
