--- src/main/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsTool.kt
@@ -1,6 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-class CodeChartsTool {
-    fun start() {}
-    private fun saveData() {}
-}
