--- src/main/code/de/tu_chemnitz/se/exercise/persist/IConfig.kt
@@ -1,7 +1,0 @@
-package de.tu_chemnitz.se.exercise.persist
-
-import org.litote.kmongo.Id
-
-interface IConfig {
-    val _id: Id<out IConfig>
-}
