+++ src/main/code/de/tuchemnitz/se/exercise/OutputDevice.kt
@@ -1,0 +1,8 @@
+package de.tuchemnitz.se.exercise
+
+interface OutputDevice {
+    fun print(message: String)
+
+    @JvmDefault
+    fun printLine(message: String) = print("$message\n")
+}
