+++ src/main/code/de/tuchemnitz/se/exercise/codecharts/CodeChartsInputValidator.kt
@@ -1,0 +1,6 @@
+package de.tuchemnitz.se.exercise.codecharts
+
+class CodeChartsInputValidator {
+    fun showInputLine() {}
+    fun validateStringInput() {}
+}
