+++ src/main/code/de/tuchemnitz/se/exercise/codecharts/CodeChartsMatrix.kt
@@ -1,0 +1,7 @@
+package de.tuchemnitz.se.exercise.codecharts
+
+class CodeChartsMatrix {
+    fun showMatrix() {}
+    private fun showStrings() {}
+    private fun rasterizeField() {}
+}
