+++ src/main/code/de/tuchemnitz/se/exercise/codecharts/CodeChartsPicture.kt
@@ -1,0 +1,6 @@
+package de.tuchemnitz.se.exercise.codecharts
+
+class CodeChartsPicture {
+    fun showPicture() {}
+    fun hidePicture() {}
+}
