+++ src/main/code/de/tuchemnitz/se/exercise/codecharts/CodeChartsTool.kt
@@ -1,0 +1,6 @@
+package de.tuchemnitz.se.exercise.codecharts
+
+class CodeChartsTool {
+    fun start() {}
+    private fun saveData() {}
+}
