+++ src/main/code/de/tuchemnitz/se/exercise/persist/IConfig.kt
@@ -1,0 +1,7 @@
+package de.tuchemnitz.se.exercise.persist
+
+import org.litote.kmongo.Id
+
+interface IConfig {
+    val _id: Id<out IConfig>
+}
