+++ src/main/code/de/tuchemnitz/se/exercise/persist/collections/ZoomMapsConfigCollection.kt
@@ -1,0 +1,9 @@
+package de.tuchemnitz.se.exercise.persist.collections
+
+import com.mongodb.client.MongoDatabase
+import de.tu_chemnitz.se.exercise.persist.configs.ZoomMapsConfig
+import org.litote.kmongo.getCollection
+
+class ZoomMapsConfigCollection(
+    private val db: MongoDatabase
+) : AbstractCollection<ZoomMapsConfig>(collection = db.getCollection())
