--- src/test/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsInputValidatorTest.kt
@@ -1,23 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
-
-@TestInstance(TestInstance.Lifecycle.PER_METHOD)
-class CodeChartsInputValidatorTest {
-    @Test
-    fun `checks whether there are the right permissions`() {
-    }
-
-    @Test
-    fun `whether the path is accessible`() {
-    }
-
-    @Test
-    fun `check whether string is too big`() {
-    }
-
-    @Test
-    fun `check whether right functions are called`() {
-    }
-}
