--- src/test/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsMatrixTest.kt
@@ -1,15 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
-
-@TestInstance(TestInstance.Lifecycle.PER_METHOD)
-class CodeChartsMatrixTest {
-    @Test
-    fun `check whether number of strings is right`() {
-    }
-
-    @Test
-    fun `check whether matrix size is in scale`() {
-    }
-}
