--- src/test/code/de/tu_chemnitz/se/exercise/codecharts/CodeChartsStringHandlerTest.kt
@@ -1,27 +1,0 @@
-package de.tu_chemnitz.se.exercise.codecharts
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
-
-@TestInstance(TestInstance.Lifecycle.PER_METHOD)
-class CodeChartsStringHandlerTest {
-    @Test
-    fun `check whether strings are duplicate`() {
-    }
-
-    @Test
-    fun `check whether free space is enough`() {
-    }
-
-    @Test
-    fun `check whether generated strings have the right format - utf8 ascii etc-`() {
-    }
-
-    @Test
-    fun `check whether list is ordered`() {
-    }
-
-    @Test
-    fun `check whether strings are really ordered`() {
-    }
-}
