+++ src/test/code/de/tuchemnitz/se/exercise/codecharts/CodeChartsPictureTest.kt
@@ -1,0 +1,19 @@
+package de.tuchemnitz.se.exercise.codecharts
+
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_METHOD)
+class CodeChartsPictureTest {
+    @Test
+    fun `checks whether there are the right permissions`() {
+    }
+
+    @Test
+    fun `whether the path is accessible`() {
+    }
+
+    @Test
+    fun `check valid file data format`() {
+    }
+}
